soc: renesas: Add initial support for RA4M1
Add support for Renesas RA4M1 and change the name of existed support
folder for Arduino R4 into RA4M1_arduino
