+++ soc/renesas/ra/ra4m1/CMakeLists.txt
@@ -1,15 +1,14 @@
-# Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+# Copyright (c) 2024 Renesas Electronics Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_include_directories(.)
 
-zephyr_library_sources_ifdef(CONFIG_SOC_OPTION_SETTING_MEMORY
+zephyr_sources(
   soc.c
 )
 
-zephyr_linker_sources_ifdef(CONFIG_SOC_OPTION_SETTING_MEMORY
-  ROM_START
-  ${CMAKE_CURRENT_SOURCE_DIR}/opt_set_mem.ld
-)
+zephyr_linker_sources(ROM_START opt_set_mem.ld)
+
+zephyr_linker_sources(SECTIONS sections.ld)
 
 set(SOC_LINKER_SCRIPT ${ZEPHYR_BASE}/include/zephyr/arch/arm/cortex_m/scripts/linker.ld CACHE INTERNAL "")
