+++ soc/renesas/ra/ra4m1/Kconfig
@@ -1,17 +1,15 @@
-# Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+# Copyright (c) 2024 Renesas Electronics Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 config SOC_SERIES_RA4M1
 	select ARM
-	select CPU_CORTEX_M4
 	select CPU_HAS_ARM_MPU
-	select CPU_CORTEX_M_HAS_SYSTICK
-	select DYNAMIC_INTERRUPTS
-	select TIMER_READS_ITS_FREQUENCY_AT_RUNTIME
+	select CPU_CORTEX_M4
+	select HAS_RENESAS_RA_FSP
+	select CPU_CORTEX_M_HAS_DWT
+	select CPU_HAS_FPU
+	select FPU
+	select HAS_SWO
 	select XIP
 	help
 	  Enable support for Renesas RA4M1 MCU series
-
-config SOC_OPTION_SETTING_MEMORY
-	bool "Option Setting Memory"
-	default y
