+++ soc/renesas/ra/ra4m1/Kconfig.defconfig
@@ -1,9 +1,12 @@
-# Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+# Copyright (c) 2024 Renesas Electronics Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 if SOC_SERIES_RA4M1
 
 config NUM_IRQS
 	default 32
+
+config PINCTRL
+	default y
 
 endif # SOC_SERIES_RA4M1
