+++ soc/renesas/ra/ra4m1/soc.h
@@ -1,7 +1,16 @@
 /*
- * Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+ * Copyright (c) 2024 Renesas Electronics Corporation
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "../common/ra_common_soc.h"
+/**
+ * @file SoC configuration macros for the Renesas RA4M1 family MCU
+ */
+
+#ifndef ZEPHYR_SOC_RENESAS_RA4M1_SOC_H_
+#define ZEPHYR_SOC_RENESAS_RA4M1_SOC_H_
+
+#include <bsp_api.h>
+
+#endif /* ZEPHYR_SOC_RENESAS_RA4M1_SOC_H_ */
