+++ soc/renesas/ra/ra4m1_arduino/Kconfig.defconfig
@@ -1,0 +1,9 @@
+# Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+# SPDX-License-Identifier: Apache-2.0
+
+if SOC_SERIES_RA4M1_ARDUINO
+
+config NUM_IRQS
+	default 32
+
+endif # SOC_SERIES_RA4M1_ARDUINO
