+++ soc/renesas/ra/ra4m1_arduino/Kconfig.soc
@@ -1,0 +1,20 @@
+# Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_RA4M1_ARDUINO
+	bool
+	select SOC_FAMILY_RENESAS_RA
+	help
+	  Renesas RA4M1
+
+config SOC_R7FA4M1AB3CFM
+	bool
+	select SOC_SERIES_RA4M1_ARDUINO
+	help
+	  R7FA4M1AB3CFM
+
+config SOC_SERIES
+	default "ra4m1_arduino" if SOC_SERIES_RA4M1_ARDUINO
+
+config SOC
+	default "r7fa4m1ab3cfm" if SOC_R7FA4M1AB3CFM
