+++ soc/renesas/ra/ra4m1_arduino/opt_set_mem.ld
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2024 Ian Morris
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ */
+
+. = 0x400;
+FILL(0xFF)
+KEEP(*(.opt_set_mem*))
+. = 0x500;
