+++ soc/renesas/ra/ra4m1_arduino/pinctrl_soc.h
