+++ soc/renesas/ra/ra4m1_arduino/soc.h
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2023 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "../common/ra_common_soc.h"
